♻️ Update dependencies, refactor Restaurant model and replace Kohttp with Fuel
The Restaurant model's id type has been changed from Long to Int across
all usages. The HTTP client library Kohttp has been replaced with Fuel.
Also, androidx.compose.material and androidx.compose.ui libraries have
 been added to improve the UI composing.

+++ app/src/main/java/app/jopiter/restaurant/repository/PreferredRestaurantRepository.kt
@@ -26,9 +26,9 @@ class PreferredRestaurantRepository(
   private val preferredRestaurantQueries: PreferredRestaurantQueries
 ) {
 
-  val preferredRestaurant = preferredRestaurantQueries.select().asFlow().map { Restaurant.get(it.executeAsOne()) }
+  val preferredRestaurant = preferredRestaurantQueries.select().asFlow().map { Restaurant.get(it.executeAsOne().toInt()) }
 
   fun setPreferredRestaurant(preferredRestaurant: Restaurant) {
-    preferredRestaurantQueries.update(preferredRestaurant.id)
+    preferredRestaurantQueries.update(preferredRestaurant.id.toLong())
   }
 }
